testscripts: add linker test for numpy
Add a test that installs Python with Devbox and then runs a simple numpy
script. It currently fails on Linux.

There's a new test-linux-amd64 script in devbox.json that runs the test.
We have some tests that are only reproducible on a specific systems, and
Docker is the easiest way of running them in a VM.